+++ sia_storage_wasm/src/object.rs
@@ -1,7 +1,8 @@
 use wasm_bindgen::prelude::*;
 
 use crate::app_key::AppKey;
-use crate::helpers::{Slab, to_js_err};
+use crate::helpers::to_js_err;
+use crate::types::Slab;
 use crate::sealed::SealedObject;
 
 /// An object stored on the Sia network. JS holds this as an opaque handle
