+++ sia_storage_wasm/src/sdk.rs
@@ -15,7 +15,8 @@ use wasm_bindgen::prelude::*;
 use wasm_bindgen::JsCast;
 
 use crate::app_key::AppKey;
-use crate::helpers::*;
+use crate::helpers::{run_local, to_js_err};
+use crate::types::{Account, Host, ObjectEvent};
 use crate::object::PinnedObject;
 use crate::packed::PackedUpload;
 use crate::streaming::Upload;
